Final commit for this week
Our final commit this week; spent some time organizing code and commenting about the stuff that was missing. Read up in Creatures.cs to see what we missed, but we will for-sure add in the future. The OOP is done which was the important thing of this lesson, we think.